Lab2: remove unused props from instructions
